% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cli_stopifnot}
\alias{cli_stopifnot}
\title{Abort if any conditions is not TRUE}
\usage{
cli_stopifnot(
  condition,
  message = "condition is not TRUE",
  call = rlang::caller_env()
)
}
\arguments{
\item{condition}{Condition to evaluate when deciding when to error}

\item{message}{Message to include in the error.}

\item{call}{Environment within which to emit the error (defaults to caller environment).}
}
\value{
Nothing
}
\description{
Abort if any conditions is not TRUE
}
\keyword{internal}
