% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoDataSource}
\alias{geoDataSource}
\title{Create sources data object in same format as observations}
\usage{
geoDataSource(longitude = NULL, latitude = NULL, call = rlang::current_env())
}
\arguments{
\item{longitude}{the locations of the potential sources in degrees longitude.}

\item{latitude}{the locations of the potential sources in degrees latitude.}

\item{call}{the environment for any checking errors to be emitted from. Only really useful for internal use.}
}
\description{
Simple function that ensures that sources are in the correct format required by Rgeoprofile. Takes longitude and latitude as input vectors and returns these same values in list format. If no values are input then default values are used.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# John Snow cholera data
geoDataSource(WaterPumps$longitude, WaterPumps$latitude)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude =
51.5235505, alpha=1, sigma=1, tau=3)
geoDataSource(sim$longitude, sim$latitude)
\dontshow{\}) # examplesIf}
}
