% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{geoMCMC}
\alias{geoMCMC}
\title{MCMC under Rgeoprofile model}
\usage{
geoMCMC(data, params, lambda = NULL, smoothprogress = TRUE)
}
\arguments{
\item{data}{input data in the format defined by \code{\link[=geoData]{geoData()}}.}

\item{params}{input parameters in the format defined by \code{\link[=geoParams]{geoParams()}}.}

\item{lambda}{bandwidth to use in posterior smoothing. If \code{NULL} then optimal bandwidth is chosen automatically by maximum-likelihood.}

\item{smoothprogress}{whether to include a progress spinner for the smoothing stage.}
}
\description{
This function carries out the main MCMC under the Rgeoprofile model. Posterior draws are smoothed to produce a posterior surface, and converted into a geoProfile. Outputs include posterior draws of alpha and sigma under the variable-sigma model.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# John Snow cholera data
d <- geoData(Cholera$longitude, Cholera$latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude =
51.5235505, alpha=1, sigma=1, tau=3)
d <- geoData(sim$longitude, sim $latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
\dontshow{\}) # examplesIf}
}
\concept{profiling}
