% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{geoRing}
\alias{geoRing}
\title{Produces a surface based on an alternative ring-search strategy}
\usage{
geoRing(params, data, source, mcmc)
}
\arguments{
\item{params}{Parameters list in the format defined by \code{\link[=geoParams]{geoParams()}}.}

\item{data}{Data object in the format defined by \code{\link[=geoData]{geoData()}}.}

\item{source}{Potential sources object in the format defined by \code{\link[=geoDataSource]{geoDataSource()}}.}

\item{mcmc}{mcmc object of the form produced by \code{\link[=geoMCMC]{geoMCMC()}}.}
}
\description{
Produces a surface based on an alternative ring-search strategy (ie searching in an expanding radius out from the 'crimes'). The output from this function can be used with \code{\link[=geoProfile]{geoProfile()}} and \code{\link[=geoReportHitscores]{geoReportHitscores()}} to produce a map and hitscores based on this strategy.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
\donttest{
# John Snow cholera data
d <- geoData(Cholera$longitude, Cholera$latitude)
s <- geoDataSource(WaterPumps$longitude, WaterPumps$latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
surface_ring <- geoRing(params = p, data = d, source = s, mcmc = m)
gp_ring <- geoProfile(surface = surface_ring)
map <- geoPlotLeaflet(params = p,
                      data = d,
                      source = s,
                      surface = gp_ring,
                      opacity = 1)
map
}
\dontshow{\}) # examplesIf}
}
\concept{profiling}
