% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{geoShapefile}
\alias{geoShapefile}
\title{Import shapefile}
\usage{
geoShapefile(fileName = NULL)
}
\arguments{
\item{fileName}{the object to be imported. Must be one of \code{SpatialPolygonsDataFrame}, \code{SpatialLinesDataFrame} or \code{RasterLayer} if it is to be used with \code{\link[=geoMask]{geoMask()}}.}
}
\description{
This function imports spatial information in the form of \code{SpatialPolygonsDataFrame}, \code{SpatialLinesDataFrame} or \code{RasterLayer} for use with \code{\link[=geoMask]{geoMask()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# load London boroughs by default
geoShapefile()
\dontshow{\}) # examplesIf}
}
\concept{spec}
