% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geoSurface3D}
\alias{geoSurface3D}
\title{Interactive 3D plot of geoprofile or raw probabilities}
\usage{
geoSurface3D(surface, surface_type = "gp")
}
\arguments{
\item{surface}{surface to plot; either the geoprofile or posteriorSurface output by \code{\link[=geoMCMC]{geoMCMC()}}.}

\item{surface_type}{type of surface; should be either \code{"gp"} for geoprofile or \code{"prob"} for posteriorSurface.}
}
\description{
Plots an interactive 3D plot of geoprofile or posterior surface (coloured according to height). This requires the \code{plotly} package to be installed.
}
\note{
Generally speaking 3D plots are not particularly useful for static data representation. As such these should only be used in decorative or interactive formats rather than as the only way of representing data.
If you are looking to represent probability surfaces in a 2D medium such as a paper, heatmaps are usually easier to interpret.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# John Snow cholera data
d <- geoData(Cholera$longitude, Cholera$latitude)
s <- geoDataSource(WaterPumps$longitude, WaterPumps$latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p, lambda=0.05)
# raw probabilities
geoSurface3D(m$posteriorSurface, surface_type = "prob")
# geoprofile
geoSurface3D(m$geoProfile, surface_type = "gp")
\dontshow{\}) # examplesIf}
}
