% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructors.R, R/spec.R
\name{gp.data}
\alias{gp.data}
\alias{geoData}
\alias{geoDataSource}
\title{Create a new data object}
\usage{
gp.data(
  longitude,
  latitude = NULL,
  lonname = "longitude",
  latname = "latitude",
  is.source = FALSE
)

geoData(...)

geoDataSource(...)
}
\arguments{
\item{longitude}{the locations of the observed data in degrees longitude (or alternatively a \code{data.frame} or \code{list} with latitude and longitude as columns).}

\item{latitude}{the locations of the observed data in degrees latitude or \code{NULL} if \code{longitude} contains the data in full.}

\item{lonname}{the name of the longitude column if \code{longitude} contains the data in full.}

\item{latname}{the name of the latitude column if \code{longitude} contains the data in full.}

\item{is.source}{\code{TRUE} if the data is to be considered a source.}

\item{...}{arguments to pass to \code{\link[=gp.data]{gp.data()}}}
}
\value{
A \code{gp.data} dataframe containing the data.
}
\description{
Create a new data object
}
\examples{
# John Snow cholera data
gp.data(WaterPumps$longitude, WaterPumps$latitude)

# Loading all-in-one
gp.data(WaterPumps)

# Loading as source data
gp.data(WaterPumps, is.source = TRUE)
}
\concept{spec}
